Ajout des classes Box et Ellipsoid, mise à jour des états de dessin, et refactorisation des primitives

+++ src/plugin/primitive/Polygon.cpp
@@ -88,7 +88,7 @@ bool Polygon::isInside(const glm::vec3 &mousePosition)
     }
 
     bool isInside = false;
-    for (int i = 0; i < points.size(); i++)
+    for (size_t i = 0; i < points.size(); i++)
     {
         int horizontal_line_y = mousePosition.y;
